Extract timeline and layout logic from ComposeScreen
